requirements.txt
qq-botpy>=1.2.1
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
